Fix linter issues

+++ src/math.rs
@@ -20,7 +20,7 @@ pub fn check_primeness(num: u8) -> bool {
 }
 
 pub fn calculator(num1: i16, operator: u8, num2: i16) -> i16 {
-    match operator as u8 {
+    match operator {
         b'+' => num1 + num2,
         b'-' => num1 - num2,
         b'*' => num1 * num2,
